Split interfaces into multiple files
